Add module Slide
