+++ src/pages/Home.jsx
@@ -1,13 +1,8 @@
-// import { Grid, Menu, Icon, Input, Header } from "semantic-ui-react";
-// import Sidebar from "../components/Sidebar";
 import { VoiceChat } from "@mui/icons-material";
 import { SearchOutlined } from "@mui/icons-material";
 import {
-  LinearProgress,
-  Badge,
   Grid,
   Stack,
-  GlobalStyles,
   Typography,
   Paper,
   Box,
@@ -16,64 +11,19 @@ import {
   TextField,
   useTheme,
 } from "@mui/material";
-import Carousel from "react-material-ui-carousel";
 
 import Img from "@/components/Img";
-import Slide from "@/components/Slide";
 import Progress from "@/components/Progress";
 
 import { Countdown, Dot } from "@/components/HotDeals";
+import Slide from "@/components/Home/Slide";
 
 const Home = () => {
-  const theme = useTheme();
-
-  var items = [
-    {
-      name: "Light Bulbs",
-      description: "Get Best Deal with Best Quality",
-      buttonText: "UP TO 45% OFF*",
-      companyName: "Sindhu Pvt.",
-    },
-    {
-      name: "Shoes",
-      description: "Get Best Deal with Best Quality",
-      buttonText: "UP TO 30% OFF*",
-      companyName: "Mindhu Pvt.",
-    },
-    {
-      name: "LED",
-      description: "Get Best Deal with Best Quality and Quantity",
-      buttonText: "UP TO 35% OFF*",
-      companyName: "Bindhu Pvt.",
-    },
-    {
-      name: "Phone",
-      description: "Get Best Deal with Great Best Quality",
-      buttonText: "UP TO 40% OFF*",
-      companyName: "Aindhu Pvt.",
-    },
-    {
-      name: "TV",
-      description: "Get Best Deal with Big TV",
-      buttonText: "UP TO 25% OFF*",
-      companyName: "Rindhu Pvt.",
-    },
-  ];
-
   return (
     <Box
       //  mt={-5}
       mx={-1}
     >
-      <GlobalStyles
-        styles={{
-          body: {
-            // backgroundColor: "#223263",
-            // backgroundColor: (theme) => theme.palette.primary.blue,
-            // backgroundColor: "primary.blue",
-          },
-        }}
-      />
       <TextField
         size="small"
         sx={{
@@ -98,32 +48,7 @@ const Home = () => {
         sx={{ mt: 1, mb: -1 }}
         alt="Offers Banner"
       />
-      <Carousel
-        autoPlay={false}
-        indicators={false}
-        // indicatorIconButtonProps={{
-        //   style: {
-        //     // padding: "10px", // 1
-        //     color: "blue", // 3
-        //   },
-        // }}
-        // indicatorContainerProps={{
-        //   style: {
-        //     marginTop: -24, // 5
-        //   },
-        // }}
-        // IndicatorIcon={<Home/>} // Previous Example
-        // activeIndicatorIconButtonProps={{
-        //   style: {
-        //     backgroundColor: "red", // 2
-        //   },
-        // }}
-      >
-        {items.map((item, i) => (
-          <Slide key={i} item={item} />
-        ))}
-      </Carousel>
-
+      <Slide />
       <Paper
         sx={{
           mt: 1,
